revised PropagatorResamplingFresnel with more user-oriented interface moved old versions of fresnel propagators to the separate module (lib_compat) updated tests
